backend/requirements.txt
Django==3.2.3
djangorestframework==3.12.4
djoser==2.1.0
django-cors-headers==3.13.0
python-dotenv==1.0.1
Pillow==9.3.0
drf-extra-fields==3.2.1
django-filter==23.1
